refactor: split LLM into package (llm/base.py + llm/gemini.py); re-export public API; remove legacy llm.py
