+++ llm/__init__.py
@@ -1,0 +1,12 @@
+# llm/__init__.py
+
+from .base import LLM, ChatMsg, MsgPart, build_gemini_contents
+from .gemini import GeminiLLM
+
+__all__ = [
+    "MsgPart",
+    "ChatMsg",
+    "LLM",
+    "build_gemini_contents",
+    "GeminiLLM",
+]
