+++ llm/chatgpt.py
@@ -1,0 +1,35 @@
+# class ChatGPT(LLM):
+#     prompt_name = "ChatGPT"
+
+#     def __init__(
+#         self,
+#         api_key: str | None = None,
+#         model: str = "gpt-4.1-nano",
+#         temperature: float = 0.7,
+#     ):
+#         self.api_key = api_key or os.getenv("OPENAI_API_KEY")
+#         if not self.api_key:
+#             raise ValueError(
+#                 "Missing ChatGPT API key. Set OPENAI_API_KEY or pass it explicitly."
+#             )
+#         self.client = AsyncOpenAI(api_key=api_key)
+#         self.model = model
+#         self.temperature = temperature
+#         self.history_size = 120
+
+#     async def query(self, system_prompt: str, user_prompt: str) -> str:
+#         logger.debug(f"Querying ChatGPT with '{system_prompt}' and '{user_prompt}'")
+#         response = await self.client.chat.completions.create(
+#             model=self.model,
+#             temperature=self.temperature,
+#             messages=[
+#                 {"role": "system", "content": system_prompt},
+#                 {"role": "user", "content": user_prompt},
+#             ],
+#             n=1,
+#         )
+
+#         if not response.choices:
+#             raise RuntimeError("LLM returned no choices.")
+
+#         return response.choices[0].message.content.strip()
