+++ llm/llama.py
@@ -1,0 +1,23 @@
+# class OllamaLLM(LLM):
+#     def __init__(self, base_url="http://serv:11434", model="gemma3"):
+#         self.prompt_name = model
+#         self.base_url = base_url.rstrip("/")
+#         self.model = model
+#         self.history_size = 5
+
+#     async def query(self, system: str, user: str) -> str:
+#         url = f"{self.base_url}/api/chat"
+#         payload = {
+#             "model": self.model,
+#             "messages": [
+#                 {"role": "system", "content": system},
+#                 {"role": "user", "content": user},
+#             ],
+#             "stream": False,
+#         }
+
+#         async with httpx.AsyncClient() as client:
+#             response = await client.post(url, json=payload, timeout=60)
+#             response.raise_for_status()
+#             data = response.json()
+#         return data.get("message", {}).get("content", "")
